+++ src/inmanta/const.py
@@ -34,6 +34,8 @@ class ResourceState(Enum):
 
 
 UNDEPLOYABLE_STATES = [ResourceState.undefined, ResourceState.skipped_for_undefined]
+TRANSIENT_STATES = [ResourceState.available, ResourceState.deploying]
+
 UNKNOWN_STRING = "<<undefined>>"
 
 """
@@ -160,3 +162,7 @@ SHUTDOWN_GRACE_IOLOOP = 10
 SHUTDOWN_GRACE_HARD = 15
 # Hard shutdown exit code
 EXIT_HARD = 3
+
+
+TIME_ISOFMT = "%Y-%m-%dT%H:%M:%S.%f"
+TIME_LOGFMT = "%Y-%m-%d %H:%M:%S"